# feat: enable to get internal "getResponseState" function at Node.js v24.x