--- a/src/response.ts
+++ b/src/response.ts
@@ -3,12 +3,8 @@
 
 import type { OutgoingHttpHeaders } from 'node:http'
 import { buildOutgoingHttpHeaders } from './utils'
-
-interface InternalBody {
-  source: string | Uint8Array | FormData | Blob | null
-  stream: ReadableStream
-  length: number | null
-}
+import { getResponseState } from './utils/internal'
+import type { InternalBody } from './utils/internal'
 
 const responseCache = Symbol('responseCache')
 const getResponseCache = Symbol('getResponseCache')
@@ -81,25 +77,12 @@ export class Response {
 Object.setPrototypeOf(Response, GlobalResponse)
 Object.setPrototypeOf(Response.prototype, GlobalResponse.prototype)
 
-const stateKey = Reflect.ownKeys(new GlobalResponse()).find(
-  (k) => typeof k === 'symbol' && k.toString() === 'Symbol(state)'
-) as symbol | undefined
-if (!stateKey) {
-  console.warn('Failed to find Response internal state key')
-}
-
 export function getInternalBody(
   response: Response | globalThis.Response
 ): InternalBody | undefined {
-  if (!stateKey) {
-    return
-  }
-
   if (response instanceof Response) {
     response = (response as any)[getResponseCache]()
   }
 
-  const state = (response as any)[stateKey] as { body?: InternalBody } | undefined
-
-  return (state && state.body) || undefined
+  return getResponseState(response as globalThis.Response)?.body
 }