fixes #240

With this PR, you will be able to apply optimisations to access internal data in v24 by executing the following.

```bash
$ node --import @hono/node-server/setup ./app.mjs
```

### What is this?

As pointed out by @timokoessler in #240, in Node.js v24.x, due to changes in https://github.com/nodejs/undici/commit/45dceeaf376325df0fee3de83bc3d9f1e82a2cfd, it is no longer possible to retrieve internal data using the previous node-server method.

Even when internal data cannot be retrieved, node-server continues to function normally; however, this may result in performance degradation for certain applications.


### Another way to get internal data

This change to ‘undici’ has made it more difficult to obtain than before, but there are ways to do so. If you execute the following code snippet in v24, you will see that it can be obtained.

```js
let getResponseState = null
const originalDeleteProperty = Reflect.deleteProperty
Reflect.deleteProperty = (target, prop) => {
  if (prop === 'getResponseState') {
    getResponseState = target.getResponseState
    Reflect.deleteProperty = originalDeleteProperty
  }
  return originalDeleteProperty(target, prop)
}

const res = new Response('Response Text')
console.log(getResponseState(res).body.source)
```


### The difficulty of making this code work

In Node.js, the Response class is dynamically initialised when `global.Response` is referenced.
This means that the code snippet above must be executed before `global.Response` is referenced for the first time. However, when executing `.mts` or `.mjs` files, `global.Response` is initialised during import resolution in most applications, so simply writing the code inside node-server will not work as expected.
To ensure that it works correctly, users must explicitly specify the `--import` option at runtime.

```bash
$ node --import @hono/node-server/setup ./app.mjs
```


### The option of giving up this optimisation

This optimisation was implemented in #145, but even without it, the speed in the ‘best case’ does not decrease. Therefore, I think there is an option to avoid making it as complex as in this PR and instead stop the optimisation that directly retrieves internal data to speed things up.

(However, if there is code that modifies the response headers, it will not achieve the ‘best case’ scenario and will remain a target for this optimisation, so I believe this optimisation applies to a fairly large number of applications.)