@usualoma 2025-05-11
Hi @yusukebe,

I believe that the only way to resolve this issue is to proceed as follows.

```bash
$ node --import @hono/node-server/setup ./app.mjs
```

Do you think it is necessary to optimise in this complex manner?

The `--import` option is not necessary for those who do not require optimisation, so it may be best to leave it as it is, to be used only by those who wish to enable it.