@yusukebe 2025-05-13
Hi @usualoma !

I am taking several benchmarks. I want to confirm in which cases there is a difference between using `--import @hono/node-server/setup` and not using it. Is there a difference in the Hello World app?