@usualoma 2025-05-13
Hi @yusukebe, Thank you for your comment.

We can see the difference in the following application. (The key point is `res.headers.set('x-foo', 'bar')`.)

```ts
import { serve } from '@hono/node-server'
import { Hono } from 'hono'

const app = new Hono()

app.get('/', () => {
    const res = new Response('foo')
    res.headers.set('x-foo', 'bar')
    return res
})

serve({
    fetch: app.fetch,
    port: 3003,
})
```

### with `--import @hono/node-server/setup`

```
$ docker run -v `pwd`:/app -w /app/hono -p 3003:3003 --rm -it node:24 node --import @hono/node-server/setup ./app.mjs
```

```
$ bombardier -d 10s --fasthttp http://localhost:3003/
Bombarding http://localhost:3003/ for 10s using 125 connection(s)
[====================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     48408.15    5503.67   57826.03
  Latency        2.58ms     1.42ms   138.00ms
  HTTP codes:
    1xx - 0, 2xx - 483930, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    11.03MB/s
```

### without `--import @hono/node-server/setup`

```
$ docker run -v `pwd`:/app -w /app/hono -p 3003:3003 --rm -it node:24 node ./app.mjs
```

```
$ bombardier -d 10s --fasthttp http://localhost:3003/
Bombarding http://localhost:3003/ for 10s using 125 connection(s)
[====================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     42615.06    6739.75   52799.27
  Latency        2.93ms     3.17ms   332.86ms
  HTTP codes:
    1xx - 0, 2xx - 426148, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     9.71MB/s
```