@yusukebe 2025-05-16
Hi @usualoma 

Thank you for the answer! This is a little bit difficult to decide. But it's okay to merge this PR!

I also took a benchmark. I can't say it too loudly, but... I believe the most important aspect of benchmarks is ***comparing them with other frameworks***. In practical use, the difference may not be significant, but when compared with other frameworks, it becomes important.

In the current `hono` implementation, creating the instance of `Response` and setting the headers after it is not a common case. We can use `c.header()` or `c.text()` to avoid this. In the benchmark case, we can use `c.header()` or `c.text()`.

Also, if users have problems in the real world, I think it's a good idea to use `--import`.

Can you make this PR ready and ping me?

---

The following are just references. I created an app that performs header addition using Hono, Express, and Fastify, and took benchmarks.

```ts
// Hono
app.get('/', (c) => {
  const res = new Response('hi')
  res.headers.set('x-foo', 'bar')
  return res
})
```

```js
// Express
import express from 'express'

const app = express()
const port = 3001

app.get('/', (req, res) => {
  res.set('x-foo', 'bar')
  res.send('H!')
})

app.listen(port)
```

```js
// Fastify
import Fastify from 'fastify'

const fastify = Fastify({ logger: false })

fastify.get('/', async (request, reply) => {
  reply.header('x-foo', 'bar')
  return 'H!'
})

fastify.listen({ port: 3002 })
```

The result is as follows. If you don't include `--import`, it will lose to Express.

![CleanShot 2025-05-17 at 05 20 54@2x](https://github.com/user-attachments/assets/6e008204-daf0-4654-b8db-bbff8d5f5e34)

The following is an example using `c.text()`.

```ts
app.get('/', (c) => {
  c.header('x-foo', 'bar')
  return c.text('H!')
})
```

![CleanShot 2025-05-17 at 05 34 34@2x](https://github.com/user-attachments/assets/cc190be4-87dd-474d-bae0-be51252c708b)

As expected, `c.text()` is fast.