@usualoma 2025-05-17
Hi @yusukebe, Thank you for reviewing and confirming the performance!


### Is it common for headers to be modified?

As @yusukebe  mentioned, returning `c.text('H!')` in the handler avoids changing the header in many cases. However, when using middleware (from the perspective of Hono and the application, not node-server), it is often reasonable to rewrite the header. Even  the middleware included in Hono, approaches that rewrite headers such as `secure-headers` and `etag` are adopted.


### Headers and status codes are frequently referenced.

Middleware probably mainly references (and modifies) `headers` and `status`. For these two properties, it seems possible to prevent performance degradation (creation of `globalThis.Response` objects) when referenced, as shown in the following PR.

https://github.com/usualoma/node-server/pull/3


### It may not be necessary to refer to the internal data structure.

Even if we refer to `headers` and `status`, if we can still achieve a fast response, I feel that hacks such as `getInternalBody` are unnecessary.

https://github.com/honojs/node-server/pull/242


### Current thoughts

I have been considering this PR to apply the same approach as `getInternalBody` to v24, but in order to maintain stable behavior in future versions and achieve performance that does not require the use of `--import`, I believe it would be best to adopt https://github.com/honojs/node-server/pull/242.


@yusukebe I'm sorry to present another proposal. What do you think?