@yusukebe 2025-05-19
Hi @usualoma !

#242 is great! I feel that the implementation of #242 seems to be simpler than the current implementation. I thought it could also cache the `url` and `statusText` properties, but they will not be referred to often, and the value will change depending on the implementation of `Request`. Shall we proceed with #242?